Replace old core module.
index.js
	-Replace the old main module of the package with `supervisor.js`.
