+++ test/lookup_results.js
@@ -6,7 +6,7 @@
 'use strict';
 
 var tape = require( 'tape' );
-var supervisor = require( '../supervisor' );
+var supervisor = require( '../');
 var lookupPoints = require( './lookup_points.json' );
 var util = require( 'util' );
